refactor: pooling.ts
